Remove duplicate Mapper class
This resolves a split package issue with this class
